sub_categoriesController updated

+++ app/Http/Controllers/CategoriesController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Sport;
 use App\Category;
 use App\Sub_category;
+use Auth;
 
 class CategoriesController extends Controller
 {
@@ -15,11 +16,16 @@ class CategoriesController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index($userId)
     {
+
+		if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
         $brands = Category::all();
 
-		//llamar a la vista devolciendo categories
+		return view('partials.admin.showCategories', array('userId'=>$userId));
     }
 
     /**
@@ -29,11 +35,15 @@ class CategoriesController extends Controller
      */
 
 	 //for creating a new subcategory with a form
-    public function create()
+    public function create($userId)
     {
-		$sports = Sport::all();
+		if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
 
-        //llamar a vista de crear category con array sports
+		$categories = Category::all();
+
+        return view('partials.admin.formCategory', array('userId'=>$userId));
     }
 
     /**
@@ -46,12 +56,17 @@ class CategoriesController extends Controller
 	 //it receives the data of a new category from a form and save it in the database
     public function store(Category $request)
     {
+		if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
         $category = new Category();
 		$subCategory->name = $request->input('name');
 		$subCategory->imagePath = $request->input('imagePath');
 		$subCategory->taxes = $request->input('taxes');
 
 		$category->save();
+		return redirect("user/$userId/categories");
     }
 
     /**
@@ -64,6 +79,10 @@ class CategoriesController extends Controller
 	 //it looks for subcategories where categoryid is this and call the view
     public function show($categoryId)
     {
+		if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
 		if (!$sub_categories = Sub_category::where('categoryId', $categoryId)->get()) {
 			abort(404);
 		}
@@ -77,14 +96,18 @@ class CategoriesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($categoryId)
+    public function edit($userId, $categoryId)
     {
+		if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
 		$sports = Sport::all();
 
 		if (!$category = Sub_category::find($categoryId)) {
 			abort(404);
 		}
-        //llamar a vista de editar category
+        return view('partials.admin.formCategory', array('userId'=>$userId, 'category'=>$category));
     }
 
     /**
@@ -96,6 +119,10 @@ class CategoriesController extends Controller
      */
     public function update(Category $request, $categoryId)
     {
+		if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
 		if (!$category = Category::find($categoryId)) {
 			abort(404);
 		}
@@ -105,6 +132,7 @@ class CategoriesController extends Controller
 		$subCategory->taxes = $request->input('taxes');
 
 		$category->save();
+		return redirect("user/$userId/categories");
     }
 
     /**
@@ -113,31 +141,25 @@ class CategoriesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($categoryId)
+    public function destroy($userId, $categoryId)
     {
+		if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
 		if (!$category = Category::find($categoryId)) {
 			abort(404);
 		}
 
-		$sub_categories = Sub_category::where('categoryId', $categoryId)->get());
-
-		/*
-			DELETE FROM products
-			WHERE products.id IN (
-	            SELECT P.id
-	            FROM (SELECT * FROM products) AS P, sub_categories S, categories C
-				WHERE P.subCategoryId=S.id
-				AND S.categoryId = 2
-	    	);
-		*/
-		$products = Products::whereIn('id', function($query) {
-		    $query->select('paper_type_id')
-		    ->from(with(new ProductCategory)->getTable())
-		    ->whereIn('category_id', ['223', '15'])
-		    ->where('active', 1);
-		})->get();
+		$sub_categories = Sub_category::where('categoryId', $categoryId)->get();
+		if (count($sub_categories)>0) {
+			$errors = array();
+			$errors[0] = "This category contains sub categories and can not be delete";
+			return view('partials.admin.showCategories', array('userId'=>$userId, 'categoryId'=>$categoryId, 'errors'=>$errors));
+		}
 
 
 		$category->delete();
+		return redirect("user/$userId/categories");
     }
 }
